Moved overflowing arithmetic functions into a separate source file
