+++ Sources/Complex/ComplexArithmetic.swift
@@ -184,27 +184,6 @@ extension Complex where Scalar: FixedWidthInteger {
     }
 }
 
-extension Complex where Scalar: FixedWidthInteger {
-
-    @_transparent
-    public func addingReportingOverflow(_ rhs: Complex<Scalar>) -> (partialValue: Complex<Scalar>, overflow: Bool) {
-        let real = self.real.addingReportingOverflow(rhs.real)
-        let imaginary = self.imaginary.addingReportingOverflow(rhs.imaginary)
-        let overflow = real.overflow || imaginary.overflow
-
-        return (partialValue: Complex<Scalar>(real: real.partialValue, imaginary: imaginary.partialValue), overflow: overflow)
-    }
-
-    @_transparent
-    public func subtractingReportingOverflow(_ rhs: Complex<Scalar>) -> (partialValue: Complex<Scalar>, overflow: Bool) {
-        let real = self.real.subtractingReportingOverflow(rhs.real)
-        let imaginary = self.imaginary.subtractingReportingOverflow(rhs.imaginary)
-        let overflow = real.overflow || imaginary.overflow
-
-        return (partialValue: Complex<Scalar>(real: real.partialValue, imaginary: imaginary.partialValue), overflow: overflow)
-    }
-}
-
 // MARK: - Multiplication
 
 extension Complex {
@@ -294,25 +273,6 @@ extension Complex where Scalar: FixedWidthInteger {
     @_transparent
     public static func .&*= (lhs: inout Complex<Scalar>, rhs: Complex<Scalar>) {
         lhs = lhs .&* rhs
-    }
-}
-
-extension Complex where Scalar: FixedWidthInteger {
-
-    @_transparent
-    public func componentwiseMultipliedFullWidth(by rhs: Complex<Scalar>) -> (high: Complex<Scalar>, low: Complex<Scalar.Magnitude>) {
-        let real = self.real.multipliedFullWidth(by: rhs.real)
-        let imaginary = self.imaginary.multipliedFullWidth(by: rhs.imaginary)
-
-        return (high: Complex<Scalar>(real: real.high, imaginary: imaginary.high), low: Complex<Scalar.Magnitude>(real: real.low, imaginary: imaginary.low))
-    }
-
-    @_transparent
-    public func componentwiseMultipliedReportingOverflow(by rhs: Complex<Scalar>) -> (partialValue: Complex<Scalar>, overflow: Bool) {
-        let real = self.real.multipliedReportingOverflow(by: rhs.real)
-        let imaginary = self.imaginary.multipliedReportingOverflow(by: rhs.imaginary)
-
-        return (partialValue: Complex<Scalar>(real: real.partialValue, imaginary: imaginary.partialValue), overflow: real.overflow || imaginary.overflow)
     }
 }
 
@@ -413,24 +373,5 @@ extension Complex where Scalar: FloatingPoint {
     @_transparent
     public static func ./= (lhs: inout Complex<Scalar>, rhs: Complex<Scalar>) {
         lhs = lhs ./ rhs
-    }
-}
-
-extension Complex where Scalar: FixedWidthInteger {
-
-    @_transparent
-    public func componentwiseDividedReportingOverflow(by rhs: Complex<Scalar>) -> (partialValue: Complex<Scalar>, overflow: Bool) {
-        let real = self.real.dividedReportingOverflow(by: rhs.real)
-        let imaginary = self.imaginary.dividedReportingOverflow(by: rhs.imaginary)
-
-        return (partialValue: Complex<Scalar>(real: real.partialValue, imaginary: imaginary.partialValue), overflow: real.overflow || imaginary.overflow)
-    }
-
-    @_transparent
-    public func componentwiseDividingFullWidth(_ dividend: (high: Complex<Scalar>, low: Complex<Scalar.Magnitude>)) -> (quotient: Complex<Scalar>, remainder: Complex<Scalar>) {
-        let real = self.real.dividingFullWidth((high: dividend.high.real, low: dividend.low.real))
-        let imaginary = self.imaginary.dividingFullWidth((high: dividend.high.imaginary, low: dividend.low.imaginary))
-
-        return (quotient: Complex<Scalar>(real: real.quotient, imaginary: imaginary.quotient), remainder: Complex<Scalar>(real: real.remainder, imaginary: imaginary.remainder))
     }
 }
